Create diceRoll.py
In this code, the roll_dice() function takes an argument num_dice, which represents the number of dice to roll. The function uses a for loop to generate a random number between 1 and 6 (inclusive) for each die, and stores the results in a list called rolls. The rolls list is then returned by the function.

In the last line of code, the number of dice is set to 2, and the roll_dice() function is called with this argument, storing the results in a variable called results. The results of the dice rolls are then printed.

You can change the number of dice to roll by changing the value of num_dice. If you want to roll one die, set num_dice to 1, if you want to roll two dice, set num_dice to 2, and so on.

Also, the code above is for a 6-sided dice, but you can easily modify it to roll dices with different sides by changing the range from 6 to the number of sides you want.