Create hangmanGame.py
This code uses a list of words to randomly choose a word for the player to guess. It creates a list of underscores the same length as the word and uses this to represent the hidden word. It then uses a while loop to run the game. The loop continues until either the player has guessed all the letters in the word or the player has made too many incorrect guesses.

The player is prompted to guess a letter and the letter is added to a list of already guessed letters. If the letter has already been guessed, the player is prompted to try again. If the letter is in the word, it is revealed in the hidden word. If the letter is not in the word, the player loses a guess.

At the end of the game, the player is informed if they won or lost and the word is revealed.

You can add more words to the word list and also you can customize the max_incorrect_guesses, the message when the player wins or loses.